# backend/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PORTAL API Documentation",
    "version": "1.0.0",
    "description": "Documentación técnica de la API de PORTAL - Sistema de Gestión de Personal y Horarios."
  },
  "servers": [{ "url": "http://localhost:4000", "description": "Servidor de Desarrollo" }],
  "components": {
    "schemas": {
...
      "AssignedShift": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "employeeId": { "type": "string" },
          "shiftPatternId": { "type": "string" },
          "startDate": { "type": "string", "pattern": "^\\d{4}-\\d{2}-\\d{2}$" },
          "endDate": { "type": "string", "nullable": true, "pattern": "^\\d{4}-\\d{2}-\\d{2}$" },
          "isDeleted": { "type": "boolean" },
          "lastModified": { "type": "number" }
        },
        "required": ["employeeId", "shiftPatternId", "startDate"]
      },
...
      "ShiftPattern": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string", "minLength": 2 },
          "cycleLengthDays": { "type": "integer", "minimum": 1 },
          "startDayOfWeek": { "type": "integer", "minimum": 0, "maximum": 6 },
          "dailySchedules": {
            "anyOf": [
              { "type": "string" },
              { "type": "array", "items": { "$ref": "#/components/schemas/PatternSchedule" } }
            ]
          },
          "color": { "type": "string" },
          "maxHoursPattern": { "type": "number" },
          "worksOnHolidays": { "type": "boolean" },
          "isDeleted": { "type": "boolean" },
          "lastModified": { "type": "number" }
        },
        "required": [
          "name",
          "cycleLengthDays",
          "startDayOfWeek",
          "dailySchedules",
          "color",
          "maxHoursPattern"
        ]
      },
      "PatternSchedule": {
        "type": "object",
        "properties": {
          "dayIndex": { "type": "integer", "minimum": 0, "maximum": 6 },
          "startTime": { "type": "string", "nullable": true },
          "endTime": { "type": "string", "nullable": true },
          "isOffDay": { "type": "boolean" },
          "hasColacion": { "type": "boolean" },
          "colacionMinutes": { "type": "number" },
          "hours": { "type": "number" }
        },
        "required": ["dayIndex", "isOffDay"]
      },
...
      "MonthlyScheduleView": {
        "type": "object",
        "properties": {
          "dateIso": { "type": "string" },
          "dayOfWeek": { "type": "string" },
          "dayOfMonth": { "type": "number" },
          "scheduleText": { "type": "string" },
          "isWorkDay": { "type": "boolean" }
        },
        "required": ["dateIso", "dayOfWeek", "dayOfMonth", "scheduleText", "isWorkDay"]
      },
      "ScheduleInfo": {
        "type": "object",
        "properties": {
          "scheduleText": { "type": "string" },
          "isWorkDay": { "type": "boolean" },
          "isHoliday": { "type": "boolean" },
          "holidayName": { "type": "string" },
          "justificationType": { "type": "string" },
          "startTime": { "type": "string" },
          "endTime": { "type": "string" },
          "hours": { "type": "number" },
          "shiftPatternId": { "type": "string" },
          "shiftPatternName": { "type": "string" },
          "patternColor": { "type": "string" },
          "hasColacion": { "type": "boolean" },
          "colacionMinutes": { "type": "number" },
          "planningStatus": { "type": "string" }
        },
        "required": ["scheduleText", "isWorkDay"]
      },
...
      "SuggestedPatternName": {
        "type": "object",
        "properties": { "suggestedName": { "type": "string" } },
        "required": ["suggestedName"]
      },
...
    },
...
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/api/shifts/patterns": {
      "get": {
        "summary": "Obtiene los patrones de turno teóricos",
        "tags": ["Shifts"],
        "parameters": [
          { "in": "query", "name": "since", "schema": { "type": "string" } },
          { "in": "query", "name": "showArchived", "schema": { "type": "boolean" } }
        ],
        "responses": {
          "200": {
            "description": "Lista de patrones",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/ShiftPattern" }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Crea un nuevo patrón de turno",
        "tags": ["Shifts"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/ShiftPattern" } }
          }
        },
        "responses": {
          "201": {
            "description": "Patrón creado",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ShiftPattern" } }
            }
          }
        }
      }
    },
    "/api/shifts/patterns/bulk": {
      "post": {
        "summary": "Carga masiva de patrones de turno",
        "tags": ["Shifts"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": { "$ref": "#/components/schemas/ShiftPattern" }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Resumen de carga",
            "content": {
              "application/json": {
                "schema": { "type": "object", "properties": { "count": { "type": "number" } } }
              }
            }
          }
        }
      }
    },
    "/api/shifts/patterns/{id}": {
      "put": {
        "summary": "Actualiza un patrón existente",
        "tags": ["Shifts"],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
        ],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/ShiftPattern" } }
          }
        },
        "responses": {
          "200": {
            "description": "Patrón actualizado",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ShiftPattern" } }
            }
          }
        }
      },
      "delete": {
        "summary": "Elimina un patrón de turno",
        "tags": ["Shifts"],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
        ],
        "security": [{ "bearerAuth": [] }],
        "responses": { "204": { "description": "Patrón eliminado" } }
      }
    },
    "/api/shifts/assignments": {
      "get": {
        "summary": "Obtiene turnos asignados a empleados",
        "tags": ["Shifts"],
        "parameters": [
          { "in": "query", "name": "page", "schema": { "type": "integer" } },
          { "in": "query", "name": "pageSize", "schema": { "type": "integer" } },
          { "in": "query", "name": "startDate", "schema": { "type": "string", "format": "date" } },
          { "in": "query", "name": "endDate", "schema": { "type": "string", "format": "date" } },
          { "in": "query", "name": "employeeId", "schema": { "type": "string" } },
          { "in": "query", "name": "since", "schema": { "type": "string" } },
          { "in": "query", "name": "showArchived", "schema": { "type": "boolean" } }
        ],
        "responses": {
          "200": {
            "description": "Lista de asignaciones",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/AssignedShift" }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "total": { "type": "integer" },
                        "page": { "type": "integer" },
                        "pageSize": { "type": "integer" },
                        "totalPages": { "type": "integer" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Asigna un turno a un empleado",
        "tags": ["Shifts"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/AssignedShift" } }
          }
        },
        "responses": {
          "201": {
            "description": "Asignación creada",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/AssignedShift" } }
            }
          }
        }
      }
    },
    "/api/shifts/assignments/bulk": {
      "post": {
        "summary": "Asignación masiva de turnos",
        "tags": ["Shifts"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": { "$ref": "#/components/schemas/AssignedShift" }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Resumen de asignaciones",
            "content": {
              "application/json": {
                "schema": { "type": "object", "properties": { "count": { "type": "number" } } }
              }
            }
          }
        }
      }
    },
    "/api/shifts/assignments/{id}": {
      "put": {
        "summary": "Actualiza una asignación de turno",
        "tags": ["Shifts"],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
        ],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/AssignedShift" } }
          }
        },
        "responses": {
          "200": {
            "description": "Asignación actualizada",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/AssignedShift" } }
            }
          }
        }
      },
      "delete": {
        "summary": "Elimina una asignación de turno",
        "tags": ["Shifts"],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }
        ],
        "security": [{ "bearerAuth": [] }],
        "responses": { "204": { "description": "Asignación eliminada" } }
      }
    },
    "/api/shifts/schedule/employee/{id}": {
      "get": {
        "summary": "Obtiene horario de un empleado para una fecha",
        "tags": ["Shifts"],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } },
          {
            "in": "query",
            "name": "date",
            "required": true,
            "schema": { "type": "string", "format": "date" }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalles del horario",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ScheduleInfo" } }
            }
          }
        }
      }
    },
    "/api/shifts/schedule/employees-on-date": {
      "get": {
        "summary": "Lista empleados programados en una fecha",
        "tags": ["Shifts"],
        "parameters": [
          {
            "in": "query",
            "name": "date",
            "required": true,
            "schema": { "type": "string", "format": "date" }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de empleados programados",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "employeeId": { "type": "string" },
                      "employeeName": { "type": "string" },
                      "shiftPatternName": { "type": "string" },
                      "startTime": { "type": "string" },
                      "endTime": { "type": "string" },
                      "patternColor": { "type": "string" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/shifts/schedule/employee/{id}/month": {
      "get": {
        "summary": "Obtiene horario mensual de un empleado",
        "tags": ["Shifts"],
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } },
          { "in": "query", "name": "year", "required": true, "schema": { "type": "integer" } },
          { "in": "query", "name": "month", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Calendario mensual",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/MonthlyScheduleView" }
                }
              }
            }
          }
        }
      }
    },
    "/api/shifts/schedule/matrix": {
      "post": {
        "summary": "Genera matriz de calendario para vista grupal",
        "tags": ["Shifts"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "startDate": { "type": "string", "format": "date" },
                  "endDate": { "type": "string", "format": "date" },
                  "employeeIds": { "type": "array", "items": { "type": "string" } }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Matriz de horarios",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "object",
                    "additionalProperties": { "$ref": "#/components/schemas/ScheduleInfo" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/shifts/monthly-plan/{employeeId}/{year}/{month}": {
      "get": {
        "summary": "Obtiene asistente de planificación mensual",
        "tags": ["Shifts"],
        "parameters": [
          { "in": "path", "name": "employeeId", "required": true, "schema": { "type": "string" } },
          { "in": "path", "name": "year", "required": true, "schema": { "type": "string" } },
          { "in": "path", "name": "month", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Planificación mensual actual",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/MonthlyScheduleView" }
                }
              }
            }
          }
        }
      }
    },
    "/api/shifts/monthly-plan": {
      "post": {
        "summary": "Guarda planificación mensual completa",
        "tags": ["Shifts"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "employeeId": { "type": "string" },
                  "month": { "type": "string" },
                  "year": { "type": "string" },
                  "patternName": { "type": "string" },
                  "dailySchedules": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "day": { "type": "integer" },
                        "type": { "type": "string", "enum": ["work", "rest"] },
                        "startTime": { "type": "string" },
                        "endTime": { "type": "string" },
                        "hours": { "type": "number" }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Resultado de la operación",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/shifts/suggest-pattern-name": {
      "get": {
        "summary": "Sugiere nombres para nuevos patrones de turno",
        "tags": ["Shifts"],
        "parameters": [
          { "in": "query", "name": "employeeId", "required": true, "schema": { "type": "string" } },
          { "in": "query", "name": "year", "required": true, "schema": { "type": "string" } },
          { "in": "query", "name": "month", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Nombre sugerido",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuggestedPatternName" }
              }
            }
          }
        }
      }
    },
    "/api/shifts/validate-conflicts": {
      "post": {
        "summary": "Valida conflictos de solapamiento de turnos",
        "tags": ["Shifts"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "employeeId": { "type": "string" },
                  "startDate": { "type": "string", "format": "date" },
                  "endDate": { "type": "string", "format": "date" },
                  "excludeAssignmentId": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Resultado de validación",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "hasConflicts": { "type": "boolean" },
                    "message": { "type": "string" },
                    "conflicts": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/AssignedShift" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}